Simplified 'derive' function
The function passed to 'derive' constraint will now take two arguments:
 - current value of the parameter
 - original value of the other parameter

It is now expected that derive function just returns a value which is directly
used for validation. This change was introduced to differentiate 'derive' from
pure custom constraints which can perform everything that derive was doing in
more-or-less similar way.